Persistent Data of Pokemon Notifications/Exclusion
In the file "../static/pokelist.json, you can list the pokemons to
notify and exclude from the map. Everytime you start the server, the
pokemon list will remain. Currently, this feature only reads data but
doesnt write. For example, modifying the list in the browswe will not
update the pokelist.json file.