Implemented using accounts from database to create worker threads
-a -p -u switches now add the account details into the db
when a new instance is started it looks for the -na flag to determine
how many accounts to pull from the db for use. default is 1 account.

ToDo:
create methods that updates accounts inuse and active flags and call
them when required
allow an existing username to be accepted as an attempt to update
password
rather than throwing unique id exception.