Database querie/update methods created
All working apart from I cant find where to place the deactivate account
method.
The error about account not being activated appears to be coming from
the api.. I cant see where tho i want to have deactivate the account and
than reboot the thread when the error occurs.
Help!!

Other than that proty much done. may need some tweeking to the in_use
timeout value

as it it you can atleast add and remove and update accounts to the db
and when a worker starts it threads the accounts for each thread are
automaticaly selected from the pool of available accounts.