step limit should be described as a search radius
since spawnpoint scanning makes use of this feature, its no longer really the number of spirals to make, its more accurately a search radius. unfortunately, its tough to explain that way because 1 step is a 70m radius, 2 steps is a 140-210m radius, 3 steps is a 280-350m radius. Because its a hex its not a fixed distance exactly.