Wrap "expire lures on Pstops during db cleanup" with a "do we parse Pstops" check
As simple as commit summary...

If we don't parse pokestops, then we don't need to expire old lures. Simple as that.

Why it bothers me? I have a huge db of pokestops (and gyms) and sadly this query runs too many times and produces no meaningful results, because they all are already lure-expired...