Major bugs
Used status when it wasn't supposed to because it's not account status but worker status, so for instance, when checking status['location'] of an account that was captcha'd an hour ago could be poiting to the current location of another account running in the same worker.
Captcha counter should be on account and account could even be put in a model for database use. Maybe next, now just getting things to work as expected.