Reverted previous commit - how captchas are counted.
The way captcha_total is calculated conflicted with this way of accounting captchas because when we retrive an account with captcha count > 0 it would double count the previous captchas.
Currently when in manual/hybrid mode captcha count for worker status doesn't go above 1 because when the account is rotated captcha count resets to 0.