Bug Fix + Speed limiter small improvement
Bug fix: Previously status['last_scanned_date'] wasn't initialized with current date. Now we use (lat, lon) to figure out if it's the first scan or not, and only check the speed limit if location is present.
Improvement: Drop search item if the worker isn't going to reach it in time - saves time wasted on a scan that would be discarded anyway.
Small fix in arg parser - uniformization.